tdf#133504 doc import: set the correct wrap
Umm, I'm amazed that
1.) this hasn't been a bigger, burning issue, and
2.) that no one has ever investigated or fixed this.
It was super easy.

The code is mind boggling though. "Auto" and "Not Beside" are NONE,
but around, none, tight, and through are "Optimal"?

How about auto is optimal, around/tight are parallel,
none is none, and through is through.

Lets try that and see how it works.

Prior patchsets submitted searched for existing examples.
They suggest that MS Word's UI only allows for notBeside and around,
but I kept the others there just in case something can trigger them.

Change-Id: I893dfd66737bcc3148b48c68cb6df778d3f528c8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/152286
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/157558
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>