sw floattable: import <w:tblOverlap w:val="never"> from DOCX
The bugdoc has 2 floating tables, the second one is shifted down in
Word, so it doesn't overlap, even if the positioning attributes would
lay them out with an overlap.

The bugdoc has <w:tblOverlap w:val="never"> set for the second table, we
can import that to Writer's
SwFormatWrapInfluenceOnObjPos::mbAllowOverlap, which was originally
added for shapes.

writerfilter/ only has access to the UNO API, so do this via the
AllowOverlap text frame property.

The layout is still missing for this, but now the doc model is correct.

(cherry picked from commit d3a0ff741f5a7ff0dcec301e5b34ee9d638acf98)

Change-Id: I25df9b75633c05af206b94ff0ad71bd240f393f5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155539
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>