tdf#144073 and tdf#147645: use bounds and angle for gradient
Passing an expanded, rotated polygon noticeably modifies the
drawing of the gradient in a slideshow due to moving of the
starting and ending colors far off the edges of the drawing
surface. So try another way and set the angle of the
gradient and draw only the unadjusted bounds.

Change-Id: I95441dfa3215396d5bc7edfa9f985335480b37de
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155729
Tested-by: Jenkins
Reviewed-by: Patrick Luby <plubius@neooffice.org>
(cherry picked from commit 9e0249b62afeca701dec27a4371f20829775422a)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/155705
Reviewed-by: Ilmari Lauhakangas <ilmari.lauhakangas@libreoffice.org>