xmloff: rename m_aObject in XMLThemeContext and clarify
Clarify what m_aObject is in the theme import (XMLThemeContext)
and rename to a better name. It is refering to an object that has
the "Theme" property, which is usually found in on XPage (master
page) or XModel, but technically could be any object that has the
property (for getting and setting the XTheme type).

Change-Id: If15a8c42a0c516cd0566f4ee9e81f7315ef1651e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156511
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
(cherry picked from commit ca120667c3855fbfc1e3579d4cf4ed0e3b93770a)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156545
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>