sw floattable, nesting: fix position of the inner follow table
The bugdoc was no longer crashing, but the inner table's follow part on
page 2 was not visible.

The problem was that lcl_ArrangeLowers() didn't try to update the fly's
position when the cell's position changed, now we do this for split
flys.

The other problem was that as SwTextFly::GetTop() is called by
SwTextFly::InitAnchoredObjList(), the inner flys were ignored while
collecting the intersecting fly frames for a paragraph, leading to an
overlap between the inner follow fly and the inner anchor text. This is
now fixed by explicitly checking for the splitfly-in-splitfly case.

With this, the ODT bugdoc now renders correctly when opened. (This is
related to tdf#55160.)

(cherry picked from commit 3aa3f0a1638a8d8006955b62bb647526768be3d8)

Change-Id: I60997e7913984872319250b0fb1cb91e02512800
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/156696
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>