oox: remove returning sRelId and rename WriteImage
Rename WriteImage -> writeGraphicToStorage to make it more clear
what the method does.

Remove returning the rel ID string in some methods that write the
graphic into the stream (or the call chain), as it is never used
later on, so it is better to not return it if it is not used (it
could be abused).

Change-Id: I78558671b82aedb34bd0f37bf7781c47ebf1cf5c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/157235
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
(cherry picked from commit 33c347fb3b0fbbae9c1005b1ca7a6105284a6d31)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/160381
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Andras Timar <andras.timar@collabora.com>