tdf#55160 sw floattable, nested DOCX imp: fix inner tables at cell start
The trouble was that in case two floating tables start at an (outer)
cell start, then everything up to the end of the first inner table ended
up in the body text, not in the outer table.

This happens because the table depth of the inserted dummy character was
incorrect.

Fix the problem by comparing what model.xml does at <w:p> start and what
OOXMLFastContextHandlerTextTable::lcl_startFastElement() did in the
dummy paragraph case, and sending the table depth in the dummy case as
well.

With this the bugdoc has the expected 19 floating tables in Writer and
is of 1 page, both matching Word.

(cherry picked from commit a33316afa4a20499159b8c900e56658512deb74a)

Change-Id: I604956f28fdc01943ab913c5aa30365376c4d2b0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/157040
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>