crashtesting: fix PDF export of fdo45193-1.docx
This started with commit 1d3d2a995239c3c71432006cb795324c56a0412a
(tdf#148687 tdf#149173 tdf#149546 sw: fix crash with textboxes,
2022-06-20), the trouble is that once anchored objects are inserted to
the anchored object list of a layout frame, the list has to be re-sorted
if something that influences the sort key changes.

Normally the way to do this is similar to commit
0a6a151c4b7c78a363fb64598fbda39db4f42d07 (Related: tdf#70062 keep
drawing anchor objects sorted, 2015-02-11), but here the problematic
call is SetFormatAttr(SwFormatAnchor(...)) in
SwTextBoxHelper::changeAnchor(), and we don't have a layout by the time
that happens, so using UpdateObjInSortedList() is not something we can
do.

In any case looks like no proper fix is coming for this soon, so at
least fix it up so we don't try to do binary search on a vector that's
not yet sorted.

This also seems to help 3 other documents from crashtesting, and also
the tdf#157571 bugdoc (related to sw floattable), which is why I started
looking at this in the first place.

(cherry picked from commit d530651e4e70febb1046727e85a8edcda610d722)

Change-Id: I9b532a39f6c28604304fa9bbecf9faf9842739a1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/157744
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>