sw floattable: add an AllowTextAfterFloatingTableBreak compat flag
Word and Writer normally only wrap the anchor text around the split fly
frame on the last page, but as
<https://bugs.documentfoundation.org/show_bug.cgi?id=61594#c33> points
out, Word has an allowTextAfterFloatingTableBreak compat flag that
allows wrapping on all pages.

Map this to a Writer compat flag for now, which affects all floating
tables in the document. The mailing list thread at
<https://lists.freedesktop.org/archives/libreoffice/2023-July/thread.html#90670>
suggests that possibly there will be demand to control this on a
per-frame level: that's not something Word supports, but we could do an
OR of the two settings in the future if wanted.

This is just the compat flag itself, layout still has to actually handle
it.

This compat flag is probably quite rare, so only write it to documents
if it's true (the default is false).

(cherry picked from commit 5b9249e950165015ba57cc2c0503381df9751bf6)

Change-Id: Ibc431f7dc11710926c115d8e03f760b7f0784304
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/157903
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>