related tdf#56258 sw sidebar SetOutline: do something without dropdown
The sidebar gained SetOutline in 7.6
with commit 3fd0b4053763aa91b0004c523e96e7d390c7b58e.

In the TOOLBAR, there is special code that tells the main button
to show the locale list outline choices - just like the dropdown does.
However, none of that code comes into play in the sidebar,
so add a special case to handle that situation.

It is easy to launch the dialog, focused on the outlinenum page.
Better yet would be to figure out how to launch the popup window
that the dropdown opens - so they do the same thing...

Change-Id: I3d6eeafaa8bdb45cfacc42e221820d664419e0e0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/159934
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Szymon Kłos <szymon.klos@collabora.com>