Avoid clipboard messages after Scheduler::ImplDeInitScheduler
When running UITests on Windows with parallelism, often there are
clipboard "content changed" messages that fail the assertion in
Scheduler::ImplDeInitScheduler. This is because the clipboard is
only uninitialized after the scheduler de-init; and even then,
the respective threads are not stopped yet.

Make sure that clipboard listeners are stopped before calling the
Scheduler::ImplDeInitScheduler in DeInitVCL.

Change-Id: I37918cdc565c7f1b3bd4f46e71c55bb5a807db6b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/160308
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>