lok calc: allow copy hyperlink outside of edit mode
Right now .uno:CopyHyperlinkLocation only works on edit mode or
inside a draw object, but LOK_CALLBACK_HYPERLINK_CLICKED is
dispatched outside of edit mode and the copy command from
the pop-up is ignored.

Since there is no text cursor, the position of the hyperlink is
needed to discriminate the correct field, since there could be
multiple hyperlink fields inside the same cell.

Change-Id: If545cf56e631adcdb914c4e799a309bc8bd0422f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/169507
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>