cool#11226 sw per-view redline on: allow both per-view and per-doc
Have two windows for a document, record changes in one view, and be
surprised that now this is not turned on in the other view since commit
ae6d396552cd3cebd7fba4942e6ca2fd5de579af (cool#11226 sw per-view redline
on: add view-aware getter, 2025-03-03).

Adding the ability to record changes per-view is intentional, but it was
not a deliberate choice to also remove the old per-document recording.

Fix this by turning the "is recording on" UI into a tristate, so the
user can enable recording for all document windows or just the current
document window -- similar to how showing changes can be off, inline and
there are also 2 more on-margin options.

The state of the new UNO commands still has to be implemented.

Change-Id: I1e5271d4cc8900a28c7a8da8d34e00fec9e2e1b0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182455
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>