cool#11226 sw per-view redline on: support this-view <-> all-views transition
Have 2 views, enable "this view" recording in view 1, then try to
transition to "all views" still in view 1, nothing happens.

This is because SwEditShell::SetRedlineFlags() checks if the redline
flags would change (for this view) and since they don't change, we
return early.

Fix the problem by passing a "record mode change" bool around, which is
set when when we keep recording on, but would switch between this-view
vs all-views to still update the various views accordingly.

Note that we have 3 states here, with the following intention:
- off: if we transition to this, all views go to off
- this view: transitioning to this from "off" just enables the current
  view, but transitioning to this from "all-views" turns off all views
- all-views: if we transition to this, all views go to on
Now all the 6 transitions between the 3 states is meant to work without
surprises.

Change-Id: I6e4d9e40f7008f680c476825a4d15dab12c6dadc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182726
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>