cool#11592 sw lok: fix sometimes lost comments on load
Load a document with the LOK API, execute
getCommandValues(".uno:ViewAnnotations"), then listen for comment
changes. If the document is complex enough, some comments won't be
included anyhwere: neither in the initial comment list, nor as a
notification later.

What happens is that SwXTextDocument::getPostIts() iterates over the
postits of the postit manager, but in case the comment is not yet laid
out (pWin is nullptr), then we just ignore that comment when creating
the initial comment list. But then SwPostItMgr::LayoutPostIts() won't
emit notifications about these comments, either -- assuming these
comments are not new. So if the LOK client is fast enough to execute
getCommandValues() and the core is slow enough with the layout of the
comments, then some initial comments in the documents are lost (they are
in the document, but they are not visible).

Fix the problem by extending SwPostItMgr::GetOrCreateAnnotationWindow()
to give info when it created the annotation window, and then in
SwPostItMgr::LayoutPostIts() notify about both freshly laid out comments
and about explicitly inserted comments.

Adding a bit of debug output confirms that now comments show up even if
we hit the unlucky "no annotation window yet" case when testing
manually:
	debug:31555:31497: SwXTextDocument::getPostIts: pWin is 0
	debug:31555:31497: SwPostItMgr::LayoutPostIts: lok notify, type is add
The first is the state right after load, the second is called by
SwPostItMgr::CalcHdl() as a user event ("on idle");

Change-Id: I32082d0014454bd492d2ae87fea42e28b639e8e5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/183999
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/187057
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>