tdf#167633 desktop: Fix dbg build crash in extend_library_path
The root cause of this problem is that the extend_library_path function
wrongly assumes that the value of the developer's LD_LIBRARY_PATH
environment variable only consists of ASCII characters and use the
rtl_uString_newFromAscii function call which is incapable of handling
non-ASCII strings.

This patch replaces the aforementioned call with rtl_string2UString,
which is capable of handling such strings.

Signed-off-by: Buo-ren Lin (OSSII) <buoren.lin@ossii.com.tw>
Change-Id: I0b42401c32e091296a389113826cac5a1f41adf4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/188155
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <stephan.bergmann@collabora.com>